backend/requirements.txt
uvicorn
fastapi
python-multipart
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
